tweak death particles, fix being stuck in the air after falling to your death bug

+++ Maps/test_map/test_map.gd
@@ -1,5 +1,5 @@
 extends Node2D
 
 func _process(_delta):
-	if %player.global_position.y > 640:
+	if %player.global_position.y > 640 and %player.isalive == true:
 		%player.kill()
